src/requirements.txt
algos==0.0.5
board==1.0
donkeycar==2.5.8
donkeycar.egg==info
eventlet==0.31.1
Flask==2.0.1
gevent_socketio==0.3.6
gym==0.18.3
h5py==2.10.0
imageio==2.9.0
keras==2.6.0
matplotlib==3.3.4
neopixel==0.0.1
numpy==1.19.5
opencv_python==4.5.3.56
pandas==1.1.5
Pillow==9.0.0
pygame==2.0.1
pytablewriter==0.64.1
PyYAML==6.0
pyzmq==22.3.0
setuptools==52.0.0.post20210125
simple_pid==1.0.1
stable_baselines==2.10.2
tensorflow==2.7.0
tensorflow_gpu==1.14.0
vae==0.1
